requirements.txt
django==5.0.4
djangorestframework==3.15.2
gunicorn==21.2.0
python-dotenv==1.0.1
whitenoise==6.7.0
Pillow==10.4.0
